CVE-2015-3152: Update cedar-14 libmysqlclient-dev from 5.5 to 5.7
The mysql 5.5/5.6 releases of libmysqlclient are still vulnerable to
CVE-2015-3152. The distro libmysqlclient-dev Ubuntu 14.04 package is
the rather old 5.5.x release, which has absolutely no way to prevent
TLS stripping when used by MySQL connectors such as MySQLdb, mysqlclient
or even Oracle's own official Connector/Python (when used in non_pure
mode).

As such (and similar to what is already done for the Postgres client),
this switches to using the official MySQL APT source, so we can get the
more up to date mysql 5.7 release of libmysqlclient.

However to prevent breakage with apps already linked against
`libmysqlclient.so.18`, we still have to install the existing mysql 5.5
based libmysqlclient18 package (which will be sourced from the
archive.ubuntu.com repo instead). This is not a problem, since
`libmysqlclient.so` will still end up being symlinked to the newer
`libmysqlclient.so.20`.

Before:
- libmysqlclient-dev  5.5.50-0ubuntu0
- libmysqlclient18    5.5.50-0ubuntu0
- mysql-common        5.5.50-0ubuntu0

After:
- libmysqlclient-dev  5.7.14-1ubuntu14.04
- libmysqlclient18    5.5.50-ubuntu0
- libmysqlclient20    5.7.14-1ubuntu14.04
- mysql-common        5.7.14-1ubuntu14.04

/usr/lib/x86_64-linux-gnu$ ls -l libmysql*
-rw-r--r-- 1 root root 5902058 Jul 12 13:37 libmysqlclient.a
lrwxrwxrwx 1 root root      20 Jul 21 12:58 libmysqlclient_r.so.18 -> libmysqlclient.so.18
lrwxrwxrwx 1 root root      24 Jul 21 12:58 libmysqlclient_r.so.18.0.0 -> libmysqlclient.so.18.0.0
lrwxrwxrwx 1 root root      20 Jul 12 13:37 libmysqlclient.so -> libmysqlclient.so.20
lrwxrwxrwx 1 root root      24 Jul 21 12:58 libmysqlclient.so.18 -> libmysqlclient.so.18.0.0
-rw-r--r-- 1 root root 3355232 Jul 21 12:59 libmysqlclient.so.18.0.0
lrwxrwxrwx 1 root root      24 Jul 12 13:37 libmysqlclient.so.20 -> libmysqlclient.so.20.3.1
-rw-r--r-- 1 root root 4283848 Jul 12 13:37 libmysqlclient.so.20.3.1
-rw-r--r-- 1 root root   18396 Jul 12 13:37 libmysqlservices.a

Fixes:
https://help.heroku.com/tickets/376633