Add libxcb.
This library is used inside of ImageMagick (included in the stack image), and its development headers are already included in the development image. Its omission in the stack is probably an error, and lead to tickets such as https://support.heroku.com/tickets/625888.

There are probably other packages we're missing. I have no idea how to exhaustively list which libs ImageMagick might try to load.