Enable ESM updates for cedar-14
Configures the ESM (Extended Security Maintenance) repository for the
cedar-14 stack image, so that it continues to receive updates now that
Ubuntu 14.04 has reached the end of its standard support lifecycle:
https://blog.ubuntu.com/2019/02/05/ubuntu-14-04-trusty-tahr

The ESM agreement requires that we not distribute images containing ESM
updates, so newly generated cedar-14 images will no longer be published
to Docker Hub.

The repository credentials will be set via the Travis dashboard, and the
APT auth config file into which they are saved is removed from the final
stack image.

Fixes:
https://gus.my.salesforce.com/a07B0000006KaKwIAK