Update postgresql-client, postgresql-server-dev and libpq to PG12
This undoes the pinning of libpq added in #148 and additionally updates
`postgresql-client` and `postgresql-server-dev` to the PG12 releases,
to support Heroku Postgres 12.

Since #148 the upstream Postgres project has improved the libpq 12 error
handling, such that it now provides a clearer error message when the
validation of connection parameters fail, which makes it easier for
customers to identify that it's due to their app having an invalid
configuration.

Release notes:
https://www.postgresql.org/docs/12/release-12.html

Refs W-7113082.