Update Heroku-20 installed packages list
Between the Ubuntu 20.04 release candidate and final release, some
packages are now no longer installed, which causes CI to fail:
https://travis-ci.org/github/heroku/stack-images/jobs/678935475

The `libapt-pkg5.90` package was previously in the base Ubuntu image,
and the `ruby-did-you-mean` package was a `depends` of `libruby2.7`.

Neither are required, and Heroku-20 is currently not released, so we
do not need to explicitly install them to maintain the installed
packages list as it was.

Refs W-7489616.