Move remaining Dockerfile env vars and commands into setup.sh
So that everything is in one place, both for humans (it's otherwise easy
to think that the APT lists aren't being cleaned up when just looking
at setup.sh) and for potential future usage of shellcheck (which will
only be able to scan the scripts, not the Dockerfiles).

The `cd /` prior to the cleanup steps has also been removed, as it's not
necessary (we neither change working directory from `/`, nor does the
current working directory ever block use of `rm`).

Refs W-7496420.