python-is-python3 (#182)
This restores 'python' as an available binary, aliased to 'python3'.

We deem this the better alternative to not having it, because programs might invoke 'python', and a 'command not found' error would presumably much more confusing than a Python 3 related error (and those should be getting pretty rare now anyway). Same for anything that has a '/usr/bin/env python' or similar shebang.

Previous stacks have 'python' aliased to 'python2', but given how that's (finally) EOL, we'd only be kicking painful migrations down the road, and people upgrading stacks are already touching their code base anyway, so better to rip the bandaid off in one go.