add libwebp-dev runtime deps (#184)
We already have libwebp6 on heroku-18 and heroku-20, and libwebpdemux3 on heroku-20, as these are dependencies of other packages.

When installing libwebp-dev during a build, this will pull down libwebpdemux3 (and also libwebpmux3 on heroku-18), as these are dependencies. A program being built may now link against these additional libraries, but they're not in the runtime image.

Explicitly adding all three libraries so the setup list makes more sense to the casual observer.