Heroku-22 (#209)
* heroku-22

Almost no changes over heroku-20 right now; only bumped libwebp and sorted out a few package versions or name changes (e.g. mime-support to media-types)

To discuss: ruby removal, python2 removal, libssl1.1
To do: libpq-dev instead of pg-server-dev, sync ImageMagick policy with upstream

GUS-W-10340283

* make libheif1 and -dev explicit

it is now a dependency of some package at runtime, so add it to the list and include the headers in the build image

should probably add this to heroku-20 as well, would be useful

* drop postgresql-server-dev

libpq-dev is enough; nobody will be building actual postgres server extensions :)

this does, however, also remove LLVM and Clang, which is great for image size, but I suspect we will want to consider adding these back

* drop Ruby

the availability of it on the system often causes confusion and issues for customers expecting a full Ruby from a buildpack but where e.g. a package.json is present in the project

trying without it for now; we can always add it back

* fix area resource policy for ImageMagick

was MB, should be MP