Add libyaml-dev to Heroku-22
I was debugging a psych issue for getting the Ruby getting started guides working with Heroku-22 when I eventually realized that heroku-22 is missing libyaml-0-2. This is installed vial `libyaml-dev` which is present on the `heroku-22-build` but not `heroku-22`. This PR adds `libyaml-22`

Internal reference to the debugging thread: https://heroku.slack.com/archives/C02GZCPPV38/p1654192955601679