Stop publishing to `heroku/heroku-private` on Docker Hub
The base images are currently double-published to Docker Hub - once to
the public `heroku/heroku` repository, and then again to the private
`heroku/heroku-private` repo.

Publishing to `heroku/heroku-private` was added in:
https://github.com/heroku/base-images/pull/130

...with the explanation that it might help internally with rollback use-cases.

However:
1. We've never had cause to use this private repository for rollback in
   the 5 years since.
2. We have a robust image rollout and rollback process, which doesn't use
   these images, but instead Cheverny (for the S3 image releases) and image
   tags on the `heroku/heroku` repo for the Docker image side:
   https://github.com/heroku/cheverny/blob/main/docs/base-image-release.md#rollback-process
3. The author of the PR that added the publishing of these images has
   confirmed that there isn't some other reason for keeping them:
   https://salesforce-internal.slack.com/archives/C01068P24S3/p1707346168976339

As such, we can stop publishing to this repository to simplify the
publishing process (which is going to be getting a lot more complex
for multi-arch support soon).

GUS-W-15018626.