Test `.img` file generation on PRs too
The base images are built as Docker images and on release, uploaded to
Docker Hub.

However, the Heroku platform currently doesn't use those images directly,
instead using a custom `.img` file from S3.

Previously this `.img` file generation was only performed on release,
meaning that the CI for PRs and on `main` wouldn't test that codepath.

Now, the generation (but not S3 upload) is performed for PRs / `main` too.

This will allow us to more easily iterate on optimising the image generation
in later PRs.

GUS-W-15244492.