Test `.img` file generation in CI on PRs and `main` too (#261)
The base images are built as Docker images and on release, uploaded to
Docker Hub. However, the Heroku platform currently doesn't use those
images directly, instead using a custom `.img` file from S3.

Previously this `.img` file generation was only performed on release,
meaning that the CI for PRs and on `main` wouldn't test that code-path,
so we might not find out about breakage until time of release.

Now, the generation (but not S3 upload) is performed for PRs/`main` too.

This will allow us to more easily iterate on optimising the image
generation in later PRs.

GUS-W-15244492.