Heroku-24: Remove gcc, make and libc6-dev from run image
GCC was added to our run images back in #127 in order to support
Ruby 2.6's then new MJIT feature:
https://www.ruby-lang.org/en/news/2018/12/25/ruby-2-6-0-released/

However, since then:
- The Ruby MJIT feature hasn't really resulted in significant
  performance benefits for real world use-cases like a Rails app.
- Ruby's MJIT has since been superseded by YJIT, which is faster and
  doesn't need GCC at runtime:
  https://shopify.engineering/yjit-just-in-time-compiler-cruby
  https://shopify.engineering/ruby-yjit-is-production-ready
- The image size impact of including build tools in our run images has
  increased considerably (#127 quoted it as 84 MB, but measuring now
  it's 203 MB).
- In a CNB world, image size is much more of a concern than in the S3
  `.img` + slug model, so we need to be more selective over what
  packages we include.

As such, this removes `gcc`, `make` and `libc6-dev` from the run image
for a 203 MB saving (they are still present in the build image, hence
zero changes to `installed-packages-*.txt` for that image).

Richard (Ruby owner) has confirmed he's fine with this change.

Note: I'm intentionally not adding `binutils` back (which was a
transitive dependency), since its 15 MB cost is not worth it for the
~once a year platform operator debugging use-case.

Before:

```
-----> Size breakdown...
       heroku/heroku:24         661MB
       heroku/heroku:24-build   1.13GB
```

After:

```
-----> Size breakdown...
       heroku/heroku:24         458MB
       heroku/heroku:24-build   1.13GB
```

Towards #266.
GUS-W-15159536.