Heroku-24: Remove Python from the run image
Since:
- Python apps will (or should be) be using Python provided by the
  Python buildpack instead.
- Non-Python buildpacks/apps typically don't need Python at runtime.
- Having Python in the run image has caused confusion in support tickets
  where the Python buildpack wasn't present (such as it being
  accidentally replaced when adding second buildpack), since at runtime
  apps then fail with a less obvious `ModuleNotFound` error instead of
  `python: command not found`.
- None of our other officially supported languages (that have their own
  buildpacks) are also installed as system packages in the base image.
- Removing Python reduces the run image size by 34 MB, and in a CNB
  world image size is a much bigger concern, so we need to be more
  selective about what packages we include.
- Once Heroku-24 GAs we can't remove packages (since it will break
  backwards compatibility given stack rebasing), however, we can add
  packages - so we should err on the side of trying out removing
  packages now.

Python is still in the build image since various non-Python use-cases
need it (for example Node.js packages that use node-gyp require Python
at install time), plus several other system packages in the build image
depend on it anyway.

I've intentionally removed the `python-is-python3` package entirely
(rather than still including it in the build image), since the vast
majority of tooling will (or should be) checking for the presence of
`python3` directly (given that's the default name on Ubuntu unless the
backward compat package is installed). And for most end-user/app
use-cases we would prefer they use the Python buildpack (rather than
system Python), so a `python: command not found` will nudge them in that
direction. We can always add `python-is-python3` back later if this
turns out to be a bigger issue than expected.

Note: The classic PHP buildpack does use Python in its
`heroku-php-apache2` and `heroku-php-nginx` scripts, however, it's only
used when `realpath` doesn't exist (eg macOS), so is unused on Heroku.
The buildpack will need to adjust for the `python-is-python3` removal,
but arguably should have done that previously (given during the Python
2 -> 3 transition the major version of `python` changed). (If it needs
to support environments where only the command `python` exists, and not
`python3`, then it can use something like:
`PYTHON=$(which python3 || which python)`)


Before (once the other PRs are merged):

```
-----> Size breakdown...
       heroku/heroku:24         424MB
       heroku/heroku:24-build   1.11GB
```

After:

```
-----> Size breakdown...
       heroku/heroku:24         390MB  (34 MB reduction)
       heroku/heroku:24-build   1.11GB (unchanged)
```

Towards #266.
GUS-W-15159536.