Heroku-24: Migrate away from transitional packages
Several of the packages being installed were actually transitional
packages, that is, the package itself is either a complete no-op (since
the functionality of the package is no needed), or else the package only
exists to depend on another package of a new name (to support people who
are still installing the package by its old name).

For the former, we can drop the dependency entirely, and for the latter
we should use the new package name directly.

Specifically:
- `apt-transport-https`: APT supports HTTPS natively (and has since
  Ubuntu 18.04), so this package doesn't do anything:
  https://packages.ubuntu.com/focal/apt-transport-https
  https://packages.ubuntu.com/focal/all/apt-transport-https/filelist
- `dnsutils` -> `bind9-dnsutils`:
  https://packages.ubuntu.com/noble/dnsutils
  https://packages.ubuntu.com/noble/all/dnsutils/filelist
- `bind9-host`: Is part of `bind9-dnsutils`, so we don't need to
  depoend on it directly:
  https://packages.ubuntu.com/noble/bind9-dnsutils
- `telnet` -> `inetutils-telnet`:
  https://packages.ubuntu.com/noble/telnet
  https://packages.ubuntu.com/noble/all/telnet/filelist

t64 mention

Note: The `libgeoip1` -> `libgeoip1t64` rename is to make CI pass since
the package has been renamed upstream as part of the 64-bit time_t
transition:
https://lists.debian.org/debian-devel-announce/2024/02/msg00005.html