Heroku-24: Update packages list for upstream changes
Upstream changes have meant the list of installed packages has
changed (`libc6-dev` used to depend upon `libnsl-dev`, but no
longer does).

We don't need this package, so won't install it explicitly:
https://packages.ubuntu.com/noble/libnsl-dev

This updates the list of packages installed in the image to match
reality, and to make CI pass.