Heroku-24: Omit transitive deps from package list
This cleans up the packages list in `setup.sh` so that it only includes
top-level dependencies, and no longer redundantly specifies packages
that are transitive dependencies of the main packages. In addition,
some packages already in the default Ubuntu 24.04 image (since they
are in the `essentials` group) have also been omitted.

(In the past before we had the installed packages list validated by CI,
these redundant transitive deps served as a way to ensure we didn't
accidentally omit a package.)

This change is a no-op in terms of the final images (note how there
is no change to the generated `installed-packages-*.txt` lists), but
makes it easier to follow/audit the remaining packages in the list.

GUS-W-15159536.