Heroku-24: Omit transitive deps from package list (#285)
This cleans up the packages list in `setup.sh` so that it only includes
top-level dependencies, and no longer redundantly specifies packages
that are transitive dependencies of the main packages. In addition,
`coreutils` has been omitted since it's already in the base Ubuntu 24.04
image (since it's in the `essentials` group).

(In the past before we had the installed packages list validated by CI,
these redundant transitive deps served as a way to ensure we didn't
accidentally omit a package.)

This change is a no-op in terms of the final images (note how there
is no change to the generated `installed-packages-*.txt` lists), but
makes it easier to follow/audit the remaining packages in the list.

GUS-W-15159536.