Heroku-24: Switch from `libgdk-pixbuf2.0-0` to `libgdk-pixbuf-2.0-0`
Since the former is a transitional package, that already depends upon
the latter:
https://packages.ubuntu.com/noble/libgdk-pixbuf2.0-0
https://packages.ubuntu.com/noble/libgdk-pixbuf-2.0-0

The new package name doesn't depend upon `libgdk-pixbuf-xlib-2.0-0`,
however, that package is deprecated, and nothing appears to use it.

This also means we're using fewer packages from `universe` (which has
lesser support guarantees), since the old packages are in `universe`
unlike the new ones that are in `main`.

GUS-W-15159536.