Track package status in `installed-packages*.txt` (#291)
During image generation we output the list of packages present in the
image and save that list to `installed-packages*.txt` files committed to
the repo. In CI, any untracked changes to these files after image
generation has been run, results in the CI run failing. This ensures we
don't ship an image update with packages missing, which could otherwise
occur if/when transitive dependencies are removed upstream.

The list of packages is generated using `dpkg-query` and a custom format
string to ensure the output doesn't include package properties that will
frequently change (such as the version field) and cause too much churn
and resultant failures in CI.

However, `dpkg-query` (like `dpkg --list`) lists not only fully
installed packages, but also packages that were partially removed, eg by
removing without using the purge option (which leaves config files
behind). This scenario can occur both if we forget to purge when
removing a package, or in cases where we actually need the config files
left behind (such as for `ca-certificates-java`).

As such, this change now includes the package status in the output, so
we can track/enforce it via the CI check. In the common case (a status
of `installed`), we filter out the status to reduce noise, and make it
easier to spot the cases that we want reviewers to focus on, when
reviewing changes to those files.

See also:
https://manpages.ubuntu.com/manpages/jammy/en/man1/dpkg-query.1.html

GUS-W-15606675.