Heroku-24: Omit transitive deps from package install list (build image)
This is the build image equivalent of #285.

Several of the packages in the list of packages to install in the build
image are actually already transitive dependencies of other packages.

For packages that we absolutely need regardless, it makes sense to still
include these transitive deps explicitly in the list. However, some of the
transitive deps are actually packages we wouldn't choose to install if they
weren't already a dependency of something else. This change omits
such entries.

This change is a no-op in terms of the final images (note how there is no
change to the generated `installed-packages-*.txt` lists), but makes it
easier to follow/audit the remaining packages in the list.

GUS-W-15159536.