Heroku-24: Migrate away from transitional packages (build image) (#292)
This is the build image equivalent of #279.

Several packages in the package install list for the build image are
actually transitional packages - which provide backwards compatibility
for packages that have since been renamed/replaced. Instead of depending
on the old transitional package name, we should switch to the new names.

Specifically:
* [libidn11-dev](https://packages.ubuntu.com/noble/libidn11-dev)
  -> [libidn-dev](https://packages.ubuntu.com/noble/libidn-dev)
* [libldap2-dev](https://packages.ubuntu.com/noble/libldap2-dev)
  -> [libldap-dev](https://packages.ubuntu.com/noble/libldap-dev)
* [libncurses5-dev](https://packages.ubuntu.com/noble/libncurses5-dev)
  and [libncursesw5-dev](https://packages.ubuntu.com/noble/libncursesw5-dev)
  -> [libncurses-dev](https://packages.ubuntu.com/noble/libncurses-dev)

This is a no-op in terms of actual libs/headers installed in the build image.

GUS-W-15616760.