Heroku-24: Re-sync ARM64 package lists
This regenerates the ARM64 packages list to match reality after #292.

CI unfortunately didn't catch this, since it currently only validates
the AMD64 list due to (a) the limitations around Docker on Linux not
supporting the containerd snapshotter and so multi-arch images, and
(b) the fact CI currently builds both architectures on the same
machine.

Also, the reason my running the build generation script locally didn't
update the ARM64 manifests, is that when I'm image size benchmarking
I sadly have to disable the containerd snapshotter since it has broken
image size calculations. With the current build scripts for this repo,
that results in only AMD64 being generated locally.

GUS-W-15616760.