Add Heroku-24 to the README
Now that Ubuntu 24.04 is GA, and we've finished with the package
cleanups/removals, we can add Heroku-24 to the README. It's listed as
"In Development" for now (similar to the CNB builder image repo README)
since Heroku-24 itself isn't yet GA and not all buildpacks support the
new stack yet.

I've also:
- Removed mention of Heroku-18, since it EOLed over a year ago.
- Added new columns for "Supported Architectures" and "Default USER",
  given that these vary across images.
- Removed the "Base" column since (a) the table is otherwise too wide to
  fit when rendered on the repo homepage on GitHub, (b) that information
  was the least useful of everything else (and also IMO potentially
  confused as many people as it helped, given it might be confused with
  the very similar tag names in the "Image" column).
- Added explanations about the build and CNB image variants.
- Documented how to handle the non-root default users.

GUS-W-15756540.