Heroku-22: remove memory limitations from imagemagick
The size and memory limitations are not useful as they don't check the
system's limits, but are just hard-coded, see:
https://github.com/ImageMagick/ImageMagick/issues/396#issuecomment-319569255

And specially for containers, the limits are better set by the
runtime.

It would be less troublesome if you could override the settings from
this configuration file, but unfortunately you can't.

Signed-off-by: David Caro <dcaro@wikimedia.org>