Add `libtirpc3t64` to Heroku-24 packages (#357)
A recent update to the iproute2 package in noble-updates removed the dependency on libtirpc3t64, and the resulting changes to the list of installed packages for Heroku-24 are causing CI failures:
https://github.com/heroku/base-images/actions/runs/16662909221/job/47163740980

Our policy is to never remove packages, since existing code may have linked to libraries contained in these packages, or rely on them in another way. If a transient dependency like this gets dropped, we need to instead explicitly add it to the list of packages (in this case, with a comment indicating we can remove it for heroku-26).

GUS-W-19212320