Add `media-types` package to Heroku-24 run-image
Since it provides `/etc/mime.types`, which is one of the locations
used by the Python `mimetypes` stdlib module:
https://github.com/python/cpython/blob/v3.13.7/Lib/mimetypes.py#L48-L58

In Heroku-22 and older the package is included due to it being a
transitive dependency of other packages (such as system Python),
however, for Heroku-24's smaller packages list we now need to
include it explicitly.

The package is ~90KB.

Fixes #360.
GUS-W-19469378.